Added refactors to fit previous PR builds and rebased with factory_implementations

+++ data-prepper-plugins/cloudwatch-logs/src/main/java/org/opensearch/dataprepper/plugins/sink/config/ThresholdConfig.java
@@ -65,7 +65,7 @@ public class ThresholdConfig {
         return logSendInterval;
     }
 
-    public long getBackOffTime() {
+    public int getBackOffTime() {
         return backOffTime;
     }
 }
